diem_config

Module config

Structs

BaseConfig

ConsensusConfig

ExecutionConfig

LeaderReputationConfig

LoggerConfig

MempoolConfig

MetricsConfig

NodeConfig

Config pulls in configuration information from the config file. This is used to set up the nodes and configure various parameters. The config file is broken up into sections for each module so that only that module can be passed around

OnDiskStorageConfig

ParseRoleError

RemoteExecutionService

RemoteService

RocksdbConfig

Port selected RocksDB options for tuning underlying rocksdb instance of DiemDB. see https://github.com/facebook/rocksdb/blob/master/include/rocksdb/options.h for detailed explanations.

RootPath

SafetyRulesConfig

SafetyRulesTestConfig

StateSyncConfig

StorageConfig

TestConfig

Enums

ConsensusProposerType

Error

ExecutionCorrectnessService

Defines how execution correctness should be run

RoleType

SafetyRulesService

Defines how safety rules should be executed

SecureBackend

Traits

PersistableConfig

Functions

invariant