Module diem_config::generator

Convenience structs and functions for generating a random set of Diem ndoes without the genesis.blob.

Structs

ValidatorSwarm

Functions

build_seed_for_network

Convenience function that builds a PeerSet containing a single peer for testing with a fully formatted NetworkAddress containing its network identity pubkey and handshake protocol version.

validator_swarm

validator_swarm_for_testing