Function diem_config::generator::build_seed_addrs

pub fn build_seed_addrs(
    seed_config: &NetworkConfig, 
    seed_base_addr: NetworkAddress
) -> SeedAddresses

Convenience function that builds a SeedAddresses containing a single peer with a fully formatted NetworkAddress containing its network identity pubkey and handshake protocol version.