Function diem_config::generator::build_seed_for_network

pub fn build_seed_for_network(
    seed_config: &NetworkConfig, 
    seed_role: PeerRole
) -> PeerSet

Convenience function that builds a PeerSet containing a single peer for testing with a fully formatted NetworkAddress containing its network identity pubkey and handshake protocol version.