Enum diem_config::network_id::NetworkId

pub enum NetworkId {
    Validator,
    Public,
    Private(String),
}

A representation of the network being used in communication. There should only be one of each NetworkId used for a single node (except for NetworkId::Public), and handshakes should verify that the NetworkId being used is the same during a handshake, to effectively ensure communication is restricted to a network. Network should be checked that it is not the DEFAULT_NETWORK

Variants

Validator

Public

Private(String)

Implementations

impl NetworkId

pub fn vfn_network() -> NetworkId

Convenience function to specify the VFN network

pub fn is_vfn_network(&self) -> bool

pub fn is_validator_network(&self) -> bool

pub fn upstream_roles(&self, role: &RoleType) -> &'static [PeerRole]

Roles for a prioritization of relative upstreams

pub fn downstream_roles(&self, role: &RoleType) -> &'static [PeerRole]

Roles for a prioritization of relative downstreams

pub fn as_str(&self) -> &str