Module diem_config::utils

Functions

default_validator_owner_auth_key_from_name

get_available_port
Return an ephemeral, available port. On unix systems, the port returned will be in the TIME_WAIT state ensuring that the OS won’t hand out this port for some grace period. Callers should be able to bind to this port given they use SO_REUSEADDR.

get_available_port_in_multiaddr

get_genesis_txn

get_local_ip
Extracts one local non-loopback IP address, if one exists. Otherwise returns None.

validator_owner_account_from_name