diem_config

Module network_id

Structs

NetworkContext

A grouping of common information between all networking code for logging. This should greatly reduce the groupings between these given everywhere, and will allow for logging accordingly.

NodeNetworkId

An intra-node identifier for a network of a node unique for a network This extra layer on top of NetworkId mainly exists for the application-layer (e.g. mempool, state sync) to differentiate between multiple public networks that a node may belong to

Enums

NetworkId

A representation of the network being used in communication. There should only be one of each NetworkId used for a single node (except for NetworkId::Public), and handshakes should verify that the NetworkId being used is the same during a handshake, to effectively ensure communication is restricted to a network. Network should be checked that it is not the DEFAULT_NETWORK